# Importer: performance improvements (accessor decoding, Shading > Use Normal Data)