Thanks for getting my last PRs in! Here are some more performance improvements:

* The first commit is just a small edit to how accessors are decoded that tries to stay out the Python interpreter as much as possible. It only gives a few tenths of a second on big models, but eh, I'll take it...
* The second tries to reduce the time spent doing the _Shading > Use Normal Data_ calculation. Here's some measurements. The last column is just to give a lower bound.

    |Sample Model|Before (Use Normal Data) |After (Use Normal Data) | — (Smooth Shading)|
    |:---|:---|:---|:---|
    |BrainStem | 1.76s | 1.55s | 1.32s |
    |GearboxAssy | 5.55s | 3.67s | 2.35s |
    |MetalRoughSpheres | 7.03s | 5.01s | 3.43s |
    |Sponza | 4.11s | 3.04s | 2.01 |

    It should have the same intent, but it does change which polys get use_smooth, it isn't exactly the same (see the commit message). The comparison against 0.9999999 is _really_ strict (there's actually only one single-precision float greater than it and still less than 1.0). I guess it's supposed to error on the side of smooth shading though.

    By the way, do you have a really good file for testing whether _Shading > Use Normal Data_ is working correctly? I'm basically just looking for any visual difference after importing a sample model.